DEVELOPERS: remove 'board/engicam' entry
In commit [1], the only 'engicam' board config present on the 2025.02.x
branch was removed.

On the master branch the 'px30core' board is present because it was
added in commit [2] not picked on LTS branch.

So the DEVELOPERS entry for Jagan Teki that match every 'engicam' board
was not removed.

This patch removes this entry to remove the post commit hook warning.

[1] 13eb6c293e configs/engicam_imx6*: remove defconfigs, broken
[2] 6e6bd098c3 configs/engicam_px30_core_defconfig: new defconfig

Signed-off-by: Thomas Perale <thomas.perale@mind.be>
Signed-off-by: Arnout Vandecappelle <arnout@rnout.be>