package/depot-tools: fix running gclient.py
As of Buildroot commit [1] which updated
the python-httplib2 package from 0.22.0 to 0.31.0, gclient.py has
been broken due to the internal socks.py file being removed from
the httplib directory (in upstream commit [2]).

As the depot-tools package doesn't have a build step, and is only
used when generating a flutter-engine tarball, no autobuild failures
have occured.

Add a simple patch that changes `import httplib2.socks` to
`import socks` and depend on the host-python-pysocks package.

[1] https://gitlab.com/buildroot.org/buildroot/-/commit/55a80f89d7321000591de349c7083a94c0349187
[2] https://github.com/httplib2/httplib2/commit/1a6ff781792099e9e14d7beb610bd99112b17b07

Fixes:
https://gitlab.com/buildroot.org/buildroot/-/jobs/12188741354

Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>
[Julien:
 - change commit id to urls
 - add "Fixes:" link
]
Signed-off-by: Julien Olivain <ju.o@free.fr>