package/coreutils: bump version to 9.8
Release notes:
https://lists.gnu.org/archive/html/coreutils-announce/2025-01/msg00000.html
https://lists.gnu.org/archive/html/coreutils-announce/2025-04/msg00000.html
https://lists.gnu.org/archive/html/coreutils-announce/2025-09/msg00000.html

Removed patch which is included in this release.

Added tarball hashes provided by upstream.

With the bump to 9.8, the build will fail on systems that are not
Y2038, such as some uClibc configurations.

In order to preserve the previous behavior, pass --disable-year2038.
See the gnulib documentation for details [1]. Contrary to what the
option name might suggest, it doesn't really disable Y2038 support,
but only the check that the system is Y2038 compliant. So even with
--disable-year2038, if the system is Y2038 compliant (uses a 64-bit
arch, uses the musl C library, or uses the glibc C library with
BR2_TIME_BITS_64=y), tar will be Y2038 compliant.

[1] https://www.gnu.org/software/gnulib/manual/html_node/Avoiding-the-year-2038-problem.html

Signed-off-by: Bernd Kuhls <bernd@kuhls.net>
Signed-off-by: Julien Olivain <ju.o@free.fr>