package/libcamera: bump version to 0.6.0
Changes between v0.5.2 and v0.6.0:
https://git.linuxtv.org/libcamera.git/tag/?h=v0.6.0

Changes between v0.5.1 and v0.5.2:
https://git.linuxtv.org/libcamera.git/tag/?h=v0.5.2

libcamera commit [1] (included in v0.5.2) introduced a usage of the
clone3() system call, introduced Kernel 5.3 [2].

Therefore, this commit adds the dependency on 5.3 headers.

This commit also changes the toolchain used in the runtime test to
the bootlin one, which includes 5.15 linux headers.

[1] https://git.linuxtv.org/libcamera.git/commit/?id=b37e69d4348a0c3db80597dac19e801d19780d1c
[2] https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=7f192e3cd316ba58c88dfa26796cf77789dd9872

Signed-off-by: Pierre-Yves Kerbrat <pkerbrat@free.fr>
[Julien:
 - squashed the libcamera bump and the runtime test fix commits
 - add clone3() comment in Config.in
 - add "header >= 5.3" in Config.in comment
]
Signed-off-by: Julien Olivain <ju.o@free.fr>