package/kbd: fix zlib build, add optional dependencies
Buildroot commit 930660890bf8b40d9bd55cd9e202e574b53d5d49 bumped kbd to
version 2.9.0 which includes optional support for various compression
libraries. This commit adds the corresponding configure options and a
fix for build errors caused by zlib.

The build error was happening when a zlib was compiled before kbd.
The issue can be reproduced with the commands:

    cat >.config <<EOF
    BR2_aarch64=y
    BR2_TOOLCHAIN_EXTERNAL=y
    BR2_TOOLCHAIN_EXTERNAL_BOOTLIN=y
    BR2_PACKAGE_KBD=y
    BR2_PACKAGE_ZLIB=y
    EOF
    make olddefconfig
    make zlib
    make kbd

The compilation fails with the erro:

    kbdfile-zlib.c: In function 'dlopen_note':
    elf-note.h:27:30: error: 'sym_gzopen' undeclared (first use in this function); did you mean 'sym_gzopen64'?
       27 | #define DLSYM_ARG(symbol__) &sym_##symbol__, STRINGIFY(symbol__),
          |                              ^~~~

Fixes:
https://autobuild.buildroot.net/results/8ff/8ff6c3d940b68069f748f12646f7516ec86172c1/

Signed-off-by: Bernd Kuhls <bernd@kuhls.net>
[Julien:
 - add commands to reproduce the issue
 - update patch "Upstream:" link to upstream commit
]
Signed-off-by: Julien Olivain <ju.o@free.fr>