package/sysprof: remove 'debuginfod' config symbol
Commit [1] was erroneously picked because it mentioned commit [2], which
is present in the LTS branch.

In fact, the 'debuginfod' option was introduced in the sysprof package in
version 48.0 (see [3] and Buildroot commit [4]), which was not picked in the
LTS branch.

This patch reverts commit [1] and removes the option it introduced, which
does not exist in sysprof 47.2.

[1] 26fd8eb217 package/sysprof: disable debuginfod support
[2] 52c5c6050c package/elfutils: bump to version 0.181
[3] https://gitlab.gnome.org/GNOME/sysprof/-/commit/d2471d4bf571c3a7c8d743a96c9207d748d763e8
[4] e743759a52 package/sysprof: bump to version 48.0

Fixes: https://gitlab.com/buildroot.org/buildroot/-/issues/152
Signed-off-by: Thomas Perale <thomas.perale@mind.be>
Acked-by: Adrian Perez de Castro <aperez@igalia.com>