toolchain/Config.in: drop BR2_TOOLCHAIN_HAS_GCC_BUG_121567
GCC bug 121567 is actually a duplicate of GCC bug 81426, which already
existed upstream, and for which we already had
BR2_TOOLCHAIN_HAS_GCC_BUG_81246.

In addition, BR2_TOOLCHAIN_HAS_GCC_BUG_81246 is correctly defined as
affecting SuperH regardless of the optimization level. Indeed, when
BR2_TOOLCHAIN_HAS_GCC_BUG_121567 was introduced, it was thought that
only -O2 and -O3 levels are affected, but -Og, -O1 and -Ofast are also
affected, causing the python3 build to fail.

Fixes:

  https://autobuild.buildroot.net/results/aca1812415ee84eef223b0fdff88f31a39775294/ (-O1 build)

  https://autobuild.buildroot.net/results/3fd91ce3ff8727d2f9c12f6721df922c60282d70/ (-Ofast build)

  https://autobuild.buildroot.net/results/e5536dcf598a51f5006343513a76f4e223e80a55/ (-Og build)

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>