package/opus: bump version to 1.6
For release notes, see:
https://opus-codec.org/release/stable/2025/12/15/libopus-1_6.html

opus 1.6 introduced optimization for Arm using Neon intrinsics.
Those are not available in soft-float ABI. This commit disables the
usage of intrinsics in that case.

This commit also adds an upstream patch to the libopusenc package.
It is because libopusenc is using internal functions of opus, which
were renamed in upstream commit [1]. Without this patch, the
opus-tools package fail to compile with error:

    src/opusenc.c: In function ‘main’:
    /.../sysroot/usr/include/opus/opusenc.h:134:63: error: implicit declaration of function ‘__opus_check_int’; did you mean ‘opus_check_int’? [-Wimplicit-function-declaration]

[1] https://gitlab.xiph.org/xiph/opus/-/commit/7817df7908318fb3a0af2a01e4da1e25643967d0

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>